fix: number of SIDS available for imputation
In the preprocessing function,when the data frame is pivoted into a country by indicator format, if an indicator is never measured for a given country, it will be NaN. Since NaN values cannot be processed in sklearn, we can either remove the country or imute it using knn imputer for the given slice of years. Note this is proposed as a temporary fix and will be expanded upon in future work.