update: missing years count mechanism`
Now when selecting predictors and targets, we look at the amount of missing values the have in years in the range [target_year - window, target_year). The missingCount variable/constant Determines for a given indicator, how much of the SIDS under study (SIDS for which the indicator is observed at some point) have to be missing for year to count as a year where too much information is missing. When set to 1, only when all SIDS understudy are missing, will the year count as missing year in the missing_years function. So now m determines for the predictor how much unecrtainity the first level imputation introduces and mtarget decides for the target how much data is available for supervised learning. So if an indicator is measured for 36 SIDS for the 15 year window (here,36 SIDS are under study and len(SIDS)-36 > n) but is only measured for 5 or 6 SIDS at for 8 years in that winow, it's measured for only ~17% of SIDS understudy for more than have of the years in the window. Since 17% < missingCount=60%, the indicator will not be considered a valid predictor for this target year as 8 > 15/m (where m =2). But the same indicator can be a valid target for this target_year since 8 < 15/mtarget (where mtarget = 1.25) and there are atleast len(SIDS)-36 SIDS that we know it's not measured for during the target_year (SIDS that WERE NOT under study). Additionally, based on the previous commit, we check if any SIDS understudy are missing for the target_year.