update: year validity check mechanism
For an indicator, we generate the years for which it can be a target by first checking ig there are years with training data (i.e., years for which the indicator is observed for some SIDS). If there are no such years, the indicator cannot be a target for imputation. If there are training data, we check if any of the SIDS under study are missing for those years (i.e., if there are SIDS that are measured for some years but not all the years for the indicator). If there are no such SIDS, we check if there are SIDS NOT under study (i.e., SIDS never measured for the indicator). If there are no such SIDS, we check if there are years for which the indicator is never measured for all SIDS. If there are no such years, then the indicator cannot be a target for imputation since all the SIDS are measured for all the years. In all other cases, we return years which have training data in their time series window (i.e., [target_year - window, target_year]) and have enough years in that window to create predictor lag (target_year >= 1970+lag, where lag is currently 3)